Split source watcher tests
